requirements.txt
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
cryptography==36.0.1
defusedxml==0.7.1
docopt==0.6.2
Flask==2.0.2
gTTS==2.2.3
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
jaraco.context==4.1.1
Jinja2==3.0.3
Js2Py==0.71
MarkupSafe==2.0.1
more-itertools==8.12.0
MouseInfo==0.1.3
outcome==1.1.0
packaging==21.3
Pillow==9.0.0
pipwin==0.5.1
playsound==1.3.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyjokes==0.6.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.1.4
PyScreeze==0.1.28
pySmartDL==1.3.4
pyttsx3==2.6
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywin32==303
requests==2.27.1
selenium==4.1.0
six==1.16.0
slask==0.1.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sport==0.0.0
sports.py==2.0.10
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.0.2
wikipedia==1.4.0
wolframalpha==5.0.0
wsproto==1.0.0
xmltodict==0.12.0
